# docs/swagger.yaml
definitions:
  api.AuthRes:
    properties:
      exchanger_balance:
        type: string
      exchanger_flag:
        type: boolean
      status:
        description: 2 The payment status of the exchange is normal, other numbers
          are arrears or no payment
        type: integer
    type: object
  api.CheckAuthRes:
    properties:
      code:
        description: 0 success 1 wrong address other failure
        type: integer
      data:
        $ref: '#/definitions/api.AuthRes'
      msg:
        type: string
    type: object
  api.ErrRes:
    properties:
      err:
        description: Error message
        type: string
    type: object
  api.GetBlockRes:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/service.Block'
      msg:
        type: string
    type: object
  api.GetReceiptsRes:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/service.Log'
        type: array
      msg:
        type: string
    type: object
  api.GetTxRes:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/service.Transaction'
      msg:
        type: string
    type: object
...
  api.ViewBlocksRes:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/service.Block'
        type: array
      msg:
        type: string
      total:
        type: integer
    type: object
  api.ViewTxsRes:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/service.Transaction'
        type: array
      msg:
        type: string
      total:
        type: integer
    type: object
...
  api.requestErbTestRes:
    properties:
      code:
        description: 0 success 1 wrong address other failure
        type: integer
      msg:
        type: string
    type: object
...
  model.Epoch:
    properties:
      creator:
        description: Creator address, also the address of royalty income
        type: string
      dir:
        description: meta information directory URL
        type: string
      exchanger:
        description: Exchange address
        type: string
      id:
        description: period ID
        type: string
      number:
        description: Starting block height
        type: integer
      royaltyRatio:
        description: The royalty rate of the same period of SNFT, the unit is one
          ten thousandth
        type: integer
      timestamp:
        description: Starting timestamp
        type: integer
      voteWeight:
        description: Weight
        type: string
    type: object
  model.Exchanger:
    properties:
      address:
        description: Exchange address
        type: string
      balance_count:
        description: Total transaction amount, unit wei
        type: string
      block_number:
        description: The block number when created
        type: integer
      creator:
        description: Creator address
        type: string
      fee_ratio:
        description: fee rate, unit 1/10,000
        type: integer
      name:
        description: Exchange name
        type: string
      nft_count:
        description: Total NFT count
        type: integer
      timestamp:
        description: Open time
        type: integer
      tx_hash:
        description: The transaction created
        type: string
      url:
        description: Exchange URL
        type: string
    type: object
...
  service.Block:
    properties:
      difficulty:
        type: string
      extraData:
        type: string
      gasLimit:
        type: string
      gasUsed:
        type: string
      miner:
        type: string
      mixHash:
        type: string
      number:
        type: string
      parentHash:
        type: string
      receiptsRoot:
        type: string
      sha3Uncles:
        type: string
      size:
        type: string
      stateRoot:
        type: string
      timestamp:
        type: string
      transactionsRoot:
        type: string
      tx_count:
        type: integer
    type: object
...
  service.EpochsRes:
    properties:
      epochs:
        description: List of system NFT periods
        items:
          $ref: '#/definitions/model.Epoch'
        type: array
      total:
        description: The total number of NFT periods in the system
        type: integer
    type: object
  service.ErrRes:
    properties:
      err_str:
        description: Error message
        type: string
    type: object
...
  service.Log:
    properties:
      address:
        type: string
      data:
        type: string
      topics:
        type: string
      tx_hash:
        type: string
    type: object
...
  service.Transaction:
    properties:
      blockHash:
        type: string
      blockNumber:
        type: string
      from:
        type: string
      gas:
        type: string
      gasPrice:
        type: string
      hash:
        type: string
      input:
        type: string
      internal_calls:
        type: string
      internal_value_transfer:
        type: string
      nonce:
        type: string
      status:
        type: string
      timestamp:
        type: string
      to:
        type: string
      token_transfer:
        type: string
      transactionIndex:
        type: string
      value:
        type: string
    type: object
...
info:
  contact: {}
  description: Block browser back-end interface, parses data from the blockchain,
    provides information retrieval services for blocks, transactions, NFT, SNFT, NFT
    collections, and exchanges
  title: block explorer API
  version: "1.0"
paths:
...
  /block/getBlock:
    get:
      consumes:
      - application/json
      deprecated: true
      description: query block
      parameters:
      - description: block number
        in: query
        name: block_number
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.GetBlockRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrRes'
      summary: query block (new /block/{number})
      tags:
      - obsolete interface
  /block/getTransaction:
    get:
      consumes:
      - application/json
      deprecated: true
      description: query transaction
      parameters:
      - description: Transaction hash
        in: query
        name: tx_hash
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.GetTxRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrRes'
      summary: query transaction (new /transaction/{hash})
      tags:
      - obsolete interface
  /block/getTransactionLogs:
    get:
      consumes:
      - application/json
      deprecated: true
      description: query receipt
      parameters:
      - description: Transaction hash
        in: query
        name: tx_hash
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.GetReceiptsRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrRes'
      summary: query receipt (new /transaction_logs/{hash})
      tags:
      - obsolete interface
...
  /block/viewBlocks:
    get:
      consumes:
      - application/json
      deprecated: true
      description: query block list
      parameters:
      - description: page
        in: query
        name: page
        type: integer
      - description: page number
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ViewBlocksRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrRes'
      summary: query block list (new /block/page)
      tags:
      - obsolete interface
  /block/viewTransactions:
    get:
      consumes:
      - application/json
      deprecated: true
      description: query transaction list
      parameters:
      - description: Address does not distinguish transmission
        in: query
        name: address
        type: string
      - description: Block number is indistinguishable
        in: query
        name: block_number
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: page number
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ViewTxsRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrRes'
      summary: query transaction list (new /transaction/page)
      tags:
      - obsolete interface
...
  /epoch:
    get:
      consumes:
      - application/json
      description: Query the system NFT period list in reverse order of creation time
      parameters:
      - description: Page, default 1
        in: query
        name: page
        type: string
      - description: Page size, default 10
        in: query
        name: page_size
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.EpochsRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/service.ErrRes'
      summary: Query the system NFT period list
      tags:
      - system NFT period
  /epoch/{id}:
    get:
      consumes:
      - application/json
      description: specifies the ID to query the NFT period information of the system,
        including 16 collection information
      parameters:
      - description: Period ID, current means query the current period
        in: path
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Epoch'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/service.ErrRes'
      summary: Query system NFT period
      tags:
      - system NFT period
...
  /exchanger/get:
    get:
      consumes:
      - application/json
      deprecated: true
      description: Query exchanges by address
      parameters:
      - description: Exchange address
        in: query
        name: addr
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Exchanger'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/service.ErrRes'
      summary: query exchange (new /exchanger/{addr})
      tags:
      - obsolete interface
...
  /extra/checkAuth:
    get:
      consumes:
      - application/json
      deprecated: true
      description: query exchange status
      parameters:
      - description: Address
        in: query
        name: address
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.CheckAuthRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrRes'
      summary: query exchange status (new /exchanger_auth)
      tags:
      - obsolete interface
  /extra/requestErbTest:
    get:
      consumes:
      - application/json
      deprecated: true
      description: request ERB test coins
      parameters:
      - description: Address
        in: query
        name: address
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.requestErbTestRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.requestErbTestRes'
      summary: request ERB test coin (new /erb_faucet)
      tags:
      - obsolete interface
...
swagger: "2.0"